requirements.txt
# Core Framework
streamlit>=1.30.0

# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.0

# Caching and task queue
redis==5.0.1
psutil==5.9.6
celery==5.3.4

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Machine Learning
xgboost>=2.0.0
scikit-learn>=1.3.0
shap>=0.44.0

# Deep Learning for LSTM and GNN
tensorflow>=2.15.0
torch>=2.1.0
torch-geometric>=2.4.0
dgl>=0.1.3

# Model versioning and experiment tracking
mlflow>=2.8.0

# Geospatial
geoalchemy2==0.14.2
shapely>=2.0.0

# Satellite data processing
h5py>=3.10.0
netCDF4>=1.6.0

# Visualization
plotly>=5.18.0

# HTTP Requests
requests>=2.31.0
aiohttp>=3.9.0
httpx==0.25.2

# Environment & Configuration
python-dotenv>=1.0.0

# Date/Time
python-dateutil>=2.8.0

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-exporter-otlp-proto-grpc==1.21.0

# Email (for alerts)
# Note: Uses standard library smtplib, no extra package needed

# Caching
cachetools>=5.3.0

# Testing (dev dependencies)
pytest>=7.4.0
pytest-asyncio==0.21.1
hypothesis==6.92.1
# pytest-cov>=4.1.0

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
